requirements.txt
# Discord Bot
discord.py>=2.3.2

# Google Sheets API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
aiohttp>=3.9.0

# Database
SQLAlchemy>=2.0
asyncpg>=0.29.0
